[#4] Refector email and password text fields
